refactor: move validation stuff to schemas
